Simplify ingestion system: remove duplication, dead code, and merge small modules
Eliminate ~440 lines and 4 files across Rust backend and React frontend while
preserving all functionality. Key changes: merge smart_folder_classifier and
batch_routes into their parent modules, extract shared helpers (ingestion_unavailable,
require_ingestion_context, usePolling), inline single-use components, and remove
dead code (BatchFolderMode, createIngestionRequest, verbose JSDoc).
