+++ src/server/static-react/src/components/tabs/FileUploadTab.jsx
@@ -1,6 +1,5 @@
 import { useState, useCallback } from 'react'
 import { ingestionClient } from '../../api/clients/ingestionClient'
-import FileUploadMode from './upload/FileUploadMode'
 
 function FileUploadTab({ onResult }) {
   const [isDragging, setIsDragging] = useState(false)
@@ -107,17 +106,34 @@ function FileUploadTab({ onResult }) {
         </div>
       )}
 
-      <FileUploadMode
-        isDragging={isDragging}
-        selectedFile={selectedFile}
-        setSelectedFile={setSelectedFile}
-        handleDragEnter={handleDragEnter}
-        handleDragOver={handleDragOver}
-        handleDragLeave={handleDragLeave}
-        handleDrop={handleDrop}
-        handleFileSelect={handleFileSelect}
-        formatFileSize={formatFileSize}
-      />
+      <div
+        className={`border-2 border-dashed p-8 text-center transition-colors ${
+          isDragging ? 'border-primary bg-surface-secondary' : 'border-border bg-surface hover:border-primary'
+        }`}
+        onDragEnter={handleDragEnter}
+        onDragOver={handleDragOver}
+        onDragLeave={handleDragLeave}
+        onDrop={handleDrop}
+      >
+        {selectedFile ? (
+          <div className="space-y-2">
+            <p className="font-medium">{selectedFile.name}</p>
+            <p className="text-sm text-secondary">{formatFileSize(selectedFile.size)}</p>
+            <button onClick={() => setSelectedFile(null)} className="text-sm text-gruvbox-red">
+              Remove
+            </button>
+          </div>
+        ) : (
+          <div className="space-y-3">
+            <p className="text-secondary">Drop file here or click to browse</p>
+            <p className="text-xs text-tertiary">PDF, DOCX, TXT, CSV, JSON, XML</p>
+            <input type="file" id="file-upload" className="hidden" onChange={handleFileSelect} />
+            <label htmlFor="file-upload" className="btn-secondary inline-block cursor-pointer">
+              Browse
+            </label>
+          </div>
+        )}
+      </div>
 
       <div className="flex items-center justify-between">
         <label className="flex items-center gap-2 text-sm cursor-pointer">
